state of art paper
In this paper, they have only focused for faults classification. But through the our research we are enhancing the predictive maintenance. Here, before the bearing is getting in to damage it can be identified. In this case it will cause for the reducing industrial downtime. And, we planned to set this model using fog computing. In edge devices which is for the quick responses and on the other hand data is filtering and forwarding to the cloud to do the predictive maintenance.